vmbot/requirements.txt
slackbot
python-vagrant
logging
validators
awscli
